+++ familyhelper-webapi-impl/src/main/java/com/dwarfeng/familyhelper/webapi/impl/service/system/PermissionGroupResponseServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
@@ -20,7 +21,6 @@ public class PermissionGroupResponseServiceImpl implements PermissionGroupRespon
 
     private final PermissionGroupMaintainService service;
 
-    @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
     public PermissionGroupResponseServiceImpl(
             @Qualifier("rbacPermissionGroupMaintainService") PermissionGroupMaintainService service
     ) {
@@ -89,56 +89,105 @@ public class PermissionGroupResponseServiceImpl implements PermissionGroupRespon
     }
 
     @Override
+    public PagedData<PermissionGroup> nameLike(String pattern, PagingInfo pagingInfo) throws ServiceException {
+        return service.lookup(PermissionGroupMaintainService.NAME_LIKE, new Object[]{pattern}, pagingInfo);
+    }
+
+    @Override
     public DispPermissionGroup getDisp(StringIdKey key) throws ServiceException {
-        return DispPermissionGroup(service.get(key));
+        PermissionGroup permissionGroup = get(key);
+        return toDisp(permissionGroup);
     }
 
     @SuppressWarnings("DuplicatedCode")
     @Override
     public PagedData<DispPermissionGroup> idLikeDisp(String pattern, PagingInfo pagingInfo) throws ServiceException {
-        PagedData<PermissionGroup> lookup = service.lookup(PermissionGroupMaintainService.ID_LIKE,
-                new Object[]{pattern}, pagingInfo);
-        List<DispPermissionGroup> DispPermissionGroups = new ArrayList<>();
-        for (PermissionGroup permissionGroup : lookup.getData()) {
-            DispPermissionGroups.add(DispPermissionGroup(permissionGroup));
-        }
-        return new PagedData<>(lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(),
-                DispPermissionGroups);
+        PagedData<PermissionGroup> lookup = idLike(pattern, pagingInfo);
+        return toDispPagedData(lookup);
     }
 
     @SuppressWarnings("DuplicatedCode")
     @Override
-    public PagedData<DispPermissionGroup> childForParentDisp(StringIdKey parentKey, PagingInfo pagingInfo) throws ServiceException {
-        PagedData<PermissionGroup> lookup = service.lookup(PermissionGroupMaintainService.CHILD_FOR_PARENT,
-                new Object[]{parentKey}, pagingInfo);
-        List<DispPermissionGroup> DispPermissionGroups = new ArrayList<>();
-        for (PermissionGroup permissionGroup : lookup.getData()) {
-            DispPermissionGroups.add(DispPermissionGroup(permissionGroup));
-        }
-        return new PagedData<>(lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(),
-                DispPermissionGroups);
+    public PagedData<DispPermissionGroup> childForParentDisp(StringIdKey parentKey, PagingInfo pagingInfo)
+            throws ServiceException {
+        PagedData<PermissionGroup> lookup = childForParent(parentKey, pagingInfo);
+        return toDispPagedData(lookup);
     }
 
     @SuppressWarnings("DuplicatedCode")
     @Override
     public PagedData<DispPermissionGroup> childForRootDisp(PagingInfo pagingInfo) throws ServiceException {
-        PagedData<PermissionGroup> lookup = service.lookup(PermissionGroupMaintainService.CHILD_FOR_PARENT,
-                new Object[]{null}, pagingInfo);
-        List<DispPermissionGroup> DispPermissionGroups = new ArrayList<>();
-        for (PermissionGroup permissionGroup : lookup.getData()) {
-            DispPermissionGroups.add(DispPermissionGroup(permissionGroup));
-        }
-        return new PagedData<>(lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(),
-                DispPermissionGroups);
+        PagedData<PermissionGroup> lookup = childForRoot(pagingInfo);
+        return toDispPagedData(lookup);
     }
 
-    private DispPermissionGroup DispPermissionGroup(PermissionGroup permissionGroup) throws ServiceException {
-        boolean hasNoChild = service.lookup(PermissionGroupMaintainService.CHILD_FOR_PARENT,
-                new Object[]{permissionGroup.getKey()}, new PagingInfo(0, 1)).getCount() <= 0;
+    @Override
+    public PagedData<DispPermissionGroup> nameLikeDisp(String pattern, PagingInfo pagingInfo) throws ServiceException {
+        PagedData<PermissionGroup> lookup = nameLike(pattern, pagingInfo);
+        return toDispPagedData(lookup);
+    }
+
+    private DispPermissionGroup toDisp(PermissionGroup permissionGroup) throws ServiceException {
         PermissionGroup parentPermissionGroup = null;
         if (Objects.nonNull(permissionGroup.getParentKey())) {
             parentPermissionGroup = service.getIfExists(permissionGroup.getParentKey());
         }
+        boolean hasNoChild = service.lookup(PermissionGroupMaintainService.CHILD_FOR_PARENT,
+                new Object[]{permissionGroup.getKey()}, new PagingInfo(0, 1)).getCount() <= 0;
         return DispPermissionGroup.of(permissionGroup, parentPermissionGroup, hasNoChild);
+    }
+
+    private PagedData<DispPermissionGroup> toDispPagedData(PagedData<PermissionGroup> lookup) throws ServiceException {
+        List<DispPermissionGroup> DispPermissionGroups = new ArrayList<>();
+        for (PermissionGroup permissionGroup : lookup.getData()) {
+            DispPermissionGroups.add(toDisp(permissionGroup));
+        }
+        return new PagedData<>(
+                lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(), DispPermissionGroups
+        );
+    }
+
+    @Override
+    public List<StringIdKey> pathFromRoot(StringIdKey key) throws ServiceException {
+        // 获取当前的权限组。
+        PermissionGroup permissionGroup = service.get(key);
+
+        // 如果权限组没有父节点，则返回空列表。
+        if (Objects.isNull(permissionGroup.getParentKey())) {
+            return new ArrayList<>();
+        }
+
+        // 定义结果列表。
+        List<StringIdKey> result = new ArrayList<>();
+
+        // 获取权限组的父节点主键，作为当前节点主键。
+        StringIdKey anchorKey = permissionGroup.getParentKey();
+
+        // 将当前节点主键添加到结果列表中。
+        result.add(anchorKey);
+
+        // 循环获取父节点的父节点，直到父节点为空。
+        while (Objects.nonNull(anchorKey)) {
+            // 获取当前节点。
+            permissionGroup = service.get(anchorKey);
+
+            // 获取当前节点的父节点主键。
+            StringIdKey parentKey = permissionGroup.getParentKey();
+
+            // 如果当前节点没有父节点，则跳出循环。
+            if (Objects.isNull(parentKey)) {
+                break;
+            }
+
+            // 将父节点主键添加到结果列表中。
+            result.add(parentKey);
+
+            // 将父节点主键作为当前节点主键。
+            anchorKey = parentKey;
+        }
+
+        // 将结果列表反转，并返回。
+        Collections.reverse(result);
+        return result;
     }
 }